Fix Tech Scout event reporting
Issue #1717
